requirements.txt
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13 